@@ -7,7 +7,7 @@
 <MudLayout>
     <MudAppBar Elevation="0">
         <MudText Typo="Typo.h6" Class="ml-4">Blazor Chat</MudText>
-        <MudAppBarSpacer />
+       <MudSpacer />
         <LoginDisplay />
     </MudAppBar>
     <MudDrawer @bind-Open="_drawerOpen" Elevation="1">
